changes in services

+++ global-service/src/main/java/in/cdac/university/globalService/entity/GmstApplicationStatusMst.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class GmstApplicationStatusMst {
 
     @Id
-    private Long unumApplicationStatusId;
+    private Integer unumApplicationStatusId;
 
     @Id
     private Integer unumIsvalid;
